memory-hotplug: bound requested size to u32
Memory hotplug resize requests used usize, allowing a MiB value to
overflow when converted to bytes.

Use u32 for requested_size_mib throughout the API, VMM, and virtio-mem
update path. The resulting maximum of just under 4 PiB always fits in a
u64 byte count, so explicit conversion validation is unnecessary.

Document the API range and test the accepted and rejected boundaries.

+++ src/vmm/src/vmm_config/memory_hotplug.rs
@@ -101,7 +101,7 @@ impl From<&VirtioMem> for MemoryHotplugConfig {
 #[serde(deny_unknown_fields)]
 pub struct MemoryHotplugSizeUpdate {
     /// Requested size in MiB to resize the hotpluggable memory to.
-    pub requested_size_mib: usize,
+    pub requested_size_mib: u32,
 }
 
 #[cfg(test)]
